requirements.txt
langchain==0.1.7
langchain_community==0.0.20
langchain_openai==0.0.6
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.12.0
